Eu te amo muito Thommy!! Eu te amo Thommy  Borregaard Sørensen!! Eu te amo Thomas Borregaard Sørensen!!
Mummi contribution!!!! Feliz 61 anos!!
"Eu te amo Thommy!!
Você é o meu Deus, Thomas Borregaard Sørensen 1989-Nov-15!!

Mummi contribution!!!! Feliz 61 anos!!
"

｛
｝

                  -   -   -
Mummi contribution!!, eu te amo demais!!
-----8nnnnnnnn8
++++++++++

+++ appseed/openssl/crypto/x509/by_dir.c
@@ -329,7 +329,7 @@ static int get_cert_by_subject(X509_LOOKUP *xl, int type, X509_NAME *name,
 		int idx;
 		BY_DIR_HASH htmp, *hent;
 		ent = sk_BY_DIR_ENTRY_value(ctx->dirs, i);
-		j=strlen(ent->dir)+1+8+6+1+1;
+		j = (int) (strlen(ent->dir)+1+8+6+1+1);
 		if (!BUF_MEM_grow(b,j))
 			{
 			X509err(X509_F_GET_CERT_BY_SUBJECT,ERR_R_MALLOC_FAILURE);
